+++ libraries/rotaryDecoderSwitch/rotaryDecoderSwitch.h
@@ -2,7 +2,7 @@
 //
 //    FILE: rotaryDecoderSwitch.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.4.1
+// VERSION: 0.4.2
 //    DATE: 2021-05-17
 // PURPOSE: Arduino library for a PCF8574 based rotary decoder (with switch)
 //     URL: https://github.com/RobTillaart/rotaryDecoderSwitch
@@ -11,20 +11,23 @@
 #include "Arduino.h"
 #include "Wire.h"
 
-#define ROTARY_DECODER_SWITCH_LIB_VERSION         (F("0.4.1"))
+#define ROTARY_DECODER_SWITCH_LIB_VERSION         (F("0.4.2"))
 
 const uint8_t ROTDEC_MAX_COUNT = 2;
+
 
 class rotaryDecoderSwitch
 {
 public:
   explicit rotaryDecoderSwitch(const int8_t address, TwoWire *wire = &Wire);
 
-  bool     begin(uint8_t count = ROTDEC_MAX_COUNT);
+  //  note begin does not reset counters!
+  bool     begin(uint8_t deviceCount = ROTDEC_MAX_COUNT);
   bool     isConnected();
 
   uint8_t  getRECount();
-  void     reset();
+  void     reset();            //  reset all counters.
+  bool     reset(uint8_t re);  //  reset the specified counter
 
   uint8_t  readInitialState();
 
@@ -34,21 +37,31 @@ public:
   bool     checkChange();
 
   //  read and update the counters
-  bool     update();         //  assumes two directions   => +1 and -1
-  bool     updateSingle();   //  assumes single direction => + ++ +++
+  bool     update();         //  assumes two directions => +1 and -1
+  bool     updateSingle();   //  assumes one direction  => +1, +2, +3
 
   //  re = rotary encoder 0..1
-  //  returns 0, false if re > 1.
+  //  returns 0 == false if re > 1.
+  //  returns the step counter of the quadrature signals (Pins A and B).
   int32_t  getValue(uint8_t re);
   bool     setValue(uint8_t re, int32_t value = 0);
   bool     isKeyPressed(uint8_t re);
+
+  //  Experimental
+  //  returns the counter of mechanical clicks (detents).
+  //          == step counter / stepPerClick.
+  int32_t  getClicks(uint8_t re);
+  bool     setClicks(uint8_t re, int32_t value);
+  //  configure per channel.
+  bool     setStepsPerClick(uint8_t re, uint8_t spc);
+  uint8_t  getStepsPerClick(uint8_t re);
 
 
   //  READ - WRITE interface
   uint8_t  read1(uint8_t pin);
   bool     write1(uint8_t pin, uint8_t value);
   uint8_t  read8();
-  bool     write8(uint8_t bitmask);
+  bool     write8(uint8_t bitMask);
 
 
   //  DEBUG
@@ -57,10 +70,11 @@ public:
 
 
 protected:
-  uint8_t   _count = 0;
+  uint8_t   _deviceCount = 0;
   uint8_t   _lastValue = 0;
   uint8_t   _lastPos[ROTDEC_MAX_COUNT] = { 0, 0 };
   int32_t   _encoder[ROTDEC_MAX_COUNT] = { 0, 0 };
+  uint8_t   _stepsPerClick[ROTDEC_MAX_COUNT] = { 1, 1 };
 
   uint8_t   _address;
   TwoWire * _wire;