+++ libraries/rotaryDecoderSwitch/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.2] - 2026-09-16
+- fix #14,
+- add getClicks(), setClicks()
+- add setStepsPerClick(), getStepsPerClick()
+- add example rotaryDecoder_getClicks.ino
+- update readme.md
+- fix internal state after reset()
+- add reset(re) to reset a single rotary encoder.
+- reduce build-CI platforms.
+- minor edits
+
 ## [0.4.1] - 2026-01-08
 - update GitHub actions
 - minor edits